pyproject.toml
[tool.poetry]
name = "pros"
version = "0.1.0"
description = ""
authors = ["Richard Hadden <richard.hadden@oeaw.ac.at>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-neomodel = "^0.0.7"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
python-cypher = "^0.20.1"
jsonpickle = "^2.2.0"
numpy = "^1.23.5"
pandas = "^1.5.2"
flexidate = "^1.4"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}
ipython = "^8.2.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
jupyter = "^1.0.0"
httpie = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
